# producer: increase performance of sampling rate lookup