Formatting a string with version and domain when looking up sampling rate hurts performance. Instead, use a dedicated structure for key. It also reduces the pressure on the GC.

Before:

```
BenchmarkDecodeEncodeNetflow/with_encoding-12             158208              8611 ns/op            8178 B/op        130 allocs/op
BenchmarkDecodeEncodeNetflow/without_encoding-12          162832              7559 ns/op            8178 B/op        130 allocs/op
```

After:

```
BenchmarkDecodeEncodeNetflow/with_encoding-12             138242              8132 ns/op            8176 B/op        129 allocs/op
BenchmarkDecodeEncodeNetflow/without_encoding-12          162312              7306 ns/op            8176 B/op        129 allocs/op
```