@lspgn 2023-12-25
Thank you! Could you add the benchmarks?

Not seeing a major difference in perfs though:

```
BenchmarkTemplatesAdd
    templates_test.go:15: Creating 1 templates
    templates_test.go:15: Creating 100 templates
    templates_test.go:15: Creating 10000 templates
    templates_test.go:15: Creating 1000000 templates
    templates_test.go:15: Creating 6573939 templates
BenchmarkTemplatesAdd-4      	 6573939	       171.8 ns/op
BenchmarkTemplatesAddGet
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 1 templates
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 100 templates
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 10000 templates
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 1000000 templates
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 24474410 templates
BenchmarkTemplatesAddGet-4   	24474410	        47.66 ns/op
```

Vs

```
BenchmarkTemplatesAdd
    templates_test.go:15: Creating 1 templates
    templates_test.go:15: Creating 100 templates
    templates_test.go:15: Creating 10000 templates
    templates_test.go:15: Creating 1000000 templates
    templates_test.go:15: Creating 4606766 templates
    templates_test.go:15: Creating 7184278 templates
BenchmarkTemplatesAdd-4      	 7184278	       168.3 ns/op
BenchmarkTemplatesAddGet
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 1 templates
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 100 templates
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 10000 templates
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 1000000 templates
    templates_test.go:22: Adding 1000 templates
    templates_test.go:24: Getting 24551028 templates
BenchmarkTemplatesAddGet-4   	24551028	        46.35 ns/op
```