@vincentbernat 2023-12-25
Well, you get the performance increase too (~2%). Moreover, because there is an allocation less, this reduces GC pressure.