@lspgn 2023-12-25
Oh yes, it's a good improvement. Was just hoping to dive into the benchmarks to understand, if easy to add? Otherwise I'll just merge.