@vincentbernat 2023-12-25
That was an end-to-end benchmark I have in Akvorado.